toggle to show unconfirmed balance
